pdf_call/requirements.txt
fastapi[standard]
openai
uvicorn
python-dotenv
pymupdf
